requirements.txt
# документация
Sphinx>=5.3.0,<5.4.0
sphinx-rtd-theme>=1.1.1,<2.0.0

# статический анализ кода (линтеры)
isort>=5.10.1,<5.11.0
flake8>=5.0.4,<5.1.0
pylint>=2.15.4,<2.16.0
mypy>=0.982,<1.0
# автоматическое форматирование кода
black>=22.10.0,<22.11.0

# автоматические тесты
pytest>=7.1.3,<7.2.0
pytest-cov>=4.0.0,<4.1.0

# валидация данных
pydantic>=1.10.2,<1.11.0

# взаимодействие с API Telegram
python-telegram-bot>=13.14,<13.15

# клиент для работы с GraphQL
gql[requests]>=3.4.0,<3.5.0
